Dynamic layout, gfx/window & display refactor
Multiple improvements and refactors across graphics, windowing, display logic and build docs:

- README: Clarify FFmpeg requirements for GTULTRA_VIDEO builds, add troubleshooting and MSYS2 package recommendations, and note vendored DLLs are runtime-only.
- src/RtMidi.cpp: Replace VLA buffer with std::vector and use safe size casts/reinterpret_cast to avoid stack VLAs.
- bme graphics/window (bme_gfx.{c,h}, bme_win.{c,h}): Add gfx_visiblexsize and gfx_setvisiblewidth to support a visible viewport smaller than the backing surface; update gfx_flip to update only the visible region; improve init failure cleanup. Add win_setwindowsize to resize the SDL window and propagate visible width.
- Console/display layout (gconsole.{c,h}, gdisplay.c): Increase MAX_COLUMNS and add DEFAULT_COLUMNS; introduce activeScreenColumns and functions to compute layout dynamically (getLayoutColumns, getSidePanelWidth, getPatternAreaWidth, getVisibleChannelCount, mapping between visual/song/channel indices, editor visual channel setters, updatescreenlayout). Update initscreen to use a backing surface width and set visible width/window size. Many pattern/order rendering functions updated to use dynamic widths, channel counts and to add robust bounds/validity checks for patterns and indices.

These changes enable flexible screen widths, multi-channel/visual-channel support, safer buffer handling, and clearer build instructions.

+++ src/resid/sid.cpp
@@ -123,8 +123,8 @@ int SID::output(short *left, short *right)
 
 	const int range = 1 << 16;
 	const int half = range >> 1;
-	short outL = extfilt.output(0) / ((4095 * 255 >> 7) * 3 * 15 * 2 / range);
-	short outR = extfilt.output(1) / ((4095 * 255 >> 7) * 3 * 15 * 2 / range);
+	int outL = extfilt.output(0) / ((4095 * 255 >> 7) * 3 * 15 * 2 / range);
+	int outR = extfilt.output(1) / ((4095 * 255 >> 7) * 3 * 15 * 2 / range);
 	if (outL >= half) {
 		outL = half - 1;
 	}
@@ -138,8 +138,8 @@ int SID::output(short *left, short *right)
 		outR = -half;
 	}
 
-	*left = outL;
-	*right = outR;
+	*left = static_cast<short>(outL);
+	*right = static_cast<short>(outR);
 
 	return outL;
 }
@@ -163,8 +163,8 @@ int SID::output(int bits, short *left, short *right)
 		outR = -half;
 	}
 
-	*left = outL;
-	*right = outR;
+	*left = static_cast<short>(outL);
+	*right = static_cast<short>(outR);
 
 	return outL;
 }
